pyproject.toml
[project]
name = "nature-finance-rag-api"
version = "0.1.0"
description = "Nature Finance data ingestion and RAG database deployment pipeline IMF CCDR semantic search API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "asyncio>=3.4.3",
    "boto3>=1.38.41",
    "google-genai>=1.9.0",
    "litellm>=1.65.1",
    "nanoid>=2.0.0",
    "numpy>=2.2.4",
    "onnx==1.17.0",
    "openai>=1.69.0",
    "pgvector>=0.4.0",
    "playwright>=1.51.0",
    "protobuf>=6.0.0.dev0",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.25.5",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "sqlmodel>=0.0.24",
    "tenacity>=9.0.0",
    "unstructured[pdf]>=0.17.2",
]

[dependency-groups]
dev = [
    "types-requests>=2.32.0.20250328",
    "types-tqdm>=4.67.0.20250417",
]
